Fix issue with incorrect id type

+++ src/main/java/dev/bibikvlad/mastermind/persistence/dao/PlayerLastSelectedDAO.java
@@ -3,6 +3,6 @@ package dev.bibikvlad.mastermind.persistence.dao;
 import dev.bibikvlad.mastermind.exceptions.PersistenceException;
 
 public interface PlayerLastSelectedDAO {
-    boolean saveOrUpdate(int id) throws PersistenceException;
-    int getLastSelected() throws PersistenceException;
+    boolean saveOrUpdate(long id) throws PersistenceException;
+    long getLastSelected() throws PersistenceException;
 }
